# contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: DriveRAG API
  version: "0.2.0"
  description: |
    Agentic Knowledge Graph RAG system API.
    This is the contract that the backend implements and the frontend consumes.
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.dingdong-rag.com/api/v1
    description: Production
paths:
...
  /chat/stream:
    post:
      operationId: chatStream
      summary: Stream a chat response (SSE)
      description: |
        Streams agent response via Server-Sent Events.
        Event types: conversation_created, tool_call, token, citation, done, error
      tags: [Chat]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: SSE stream of chat events
          content:
            text/event-stream: {}
  /chat/message:
    post:
      operationId: chatMessage
      summary: Non-streaming chat response
      tags: [Chat]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: Complete chat message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /chat/conversations:
    get:
      operationId: listConversations
      summary: List conversations
      tags: [Chat]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConversationResponse"
    post:
      operationId: createConversation
      summary: Create a new conversation
      tags: [Chat]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationCreate"
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationResponse"
  /chat/conversations/{conversation_id}:
    get:
      operationId: getConversation
      summary: Get conversation details
      tags: [Chat]
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Conversation details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationResponse"
        "404":
          description: Conversation not found
    delete:
      operationId: deleteConversation
      summary: Delete a conversation
      tags: [Chat]
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Conversation deleted
        "404":
          description: Conversation not found
  /chat/conversations/{conversation_id}/messages:
    get:
      operationId: getMessages
      summary: Get conversation messages
      tags: [Chat]
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        "200":
          description: Message history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Conversation not found

  # ── Admin: Ingestion ──────────────────────────────────────────────────
...
components:
  schemas:
...
    ChatRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 10000
        conversation_id:
          type: string
          format: uuid
          nullable: true
        model:
          type: string
          nullable: true
          description: Model override for this request (e.g. 'anthropic/claude-sonnet-4-20250514')
        vector_weight:
          type: number
          minimum: 0
          maximum: 1
          default: 0.5
    Citation:
      type: object
      properties:
        document_id:
          type: string
          format: uuid
          nullable: true
        document_name:
          type: string
          nullable: true
        chunk_id:
          type: string
          format: uuid
          nullable: true
        entity_id:
          type: string
          format: uuid
          nullable: true
        entity_name:
          type: string
          nullable: true
        content_snippet:
          type: string
        page_number:
          type: integer
          nullable: true
        source_type:
          type: string
          enum: [vector, graph, both]
        relevance_score:
          type: number
    MessageResponse:
      type: object
      required: [id, conversation_id, role, content, created_at]
      properties:
        id:
          type: string
          format: uuid
        conversation_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        citations:
          type: array
          items:
            $ref: "#/components/schemas/Citation"
        tool_calls:
          type: array
          items:
            type: object
          nullable: true
        token_count:
          type: integer
        created_at:
          type: string
          format: date-time
    ConversationCreate:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
          default: New conversation
    ConversationResponse:
      type: object
      required: [id, title, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        message_count:
          type: integer
          default: 0

    # ── Ingestion ─────────────────────────────────────────────────────
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
...